requirements.txt
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.5
Django==3.1.7
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
pep8==1.7.1
Pillow==8.1.1
pycodestyle==2.6.0
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
